requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
Django==4.2
django-crispy-forms==2.0
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
Pygments==2.15.1
PySocks==1.7.1
pytz==2023.3
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.3.5
sqlparse==0.4.4
urllib3==1.26.15
